updated to a cleaner file hierarchy
