--- frontend/src/modules/Containers/CreateTournamentContainer.js
@@ -1,25 +1,0 @@
-import { connect } from 'react-redux';
-import {
-  addTournament,
-  addTournamentStart,
-  addTournamentSuccess,
-  addTournamentFailure
-} from '../Actions';
-import CreateTournament from '../CreateTournament';
-
-const mapStateToProps = state => {
-  return {
-    tournamentList: state.tournamentList.tournaments
-  };
-};
-
-const mapDispatchToProps = dispatch => {
-  return {
-    addTournament: data => dispatch(addTournament(data))
-  };
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(CreateTournament);
