--- frontend/src/modules/Containers/TournamentsListContainer.js
@@ -1,25 +1,0 @@
-import { connect } from 'react-redux';
-import {
-  fetchTournaments,
-  fetchTournamentsStart,
-  fetchTournamentsSuccess,
-  fetchTournamentsFailure
-} from '../Actions';
-import TournamentsList from '../TournamentsList';
-
-const mapStateToProps = state => {
-  return {
-    tournamentList: state.tournamentList.tournaments
-  };
-};
-
-const mapDispatchToProps = dispatch => {
-  return {
-    fetchTournaments: () => dispatch(fetchTournaments())
-  };
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(TournamentsList);
